requirements.txt
skpy
spotipy
python-dotenv
schedule